Add the giving to the student the final grade and allowing him to leave if he fails the test. Add a thread to the Student so that he can clean the newly genereted Uczen that are not part of the Egzam.

+++ list.c
@@ -75,7 +75,7 @@ static float normalization(float grade) {
     }
 }
 
-int calculateFinalGrade(StudentGrade* head, int studentID) {
+float calculateFinalGrade(StudentGrade* head, int studentID) {
     StudentGrade* next = head;
     while (next->studentID != studentID) {
         next = next->next;
